backend/requirements.txt
fastapi
uvicorn[standard]
pyautogen
youtube-transcript-api
python-dotenv
groq
requests
